Created Repository classes and added test function
New Classes:
+StudentRepository
+InstructorRepository

Added new Dependency in Pom.xml
Added test function in BootApplication

+++ src/main/java/Model/Person.java
@@ -7,7 +7,7 @@ public abstract class Person {
 
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private Integer id = null;
+    private Long id = null;
     private String name = null;
     private String type = null;
 
@@ -21,7 +21,7 @@ public abstract class Person {
         this.type = type;
     }
 
-    public Integer getId(){
+    public Long getId(){
         return this.id;
     }
